refactor(gateway): extract chat reply dispatch
