+++ src/gateway/server-methods/chat-send-reply-dispatch.test.ts
@@ -1,0 +1,82 @@
+import { describe, expect, it, vi } from "vitest";
+import { setReplyPayloadMetadata } from "../../auto-reply/reply-payload.js";
+import {
+  buildTranscriptReplyText,
+  createChatSendReplyDispatch,
+} from "./chat-send-reply-dispatch.js";
+
+describe("buildTranscriptReplyText", () => {
+  it("keeps reply directives and safe media while suppressing reasoning", () => {
+    expect(
+      buildTranscriptReplyText([
+        { text: "hidden", isReasoning: true },
+        {
+          text: "Hello",
+          replyToId: "message-1",
+          mediaUrls: ["https://example.test/photo.png"],
+        },
+        {
+          text: "Listen",
+          audioAsVoice: true,
+          mediaUrl: "https://example.test/clip.mp3",
+        },
+        {
+          text: "private",
+          sensitiveMedia: true,
+          mediaUrl: "https://example.test/private.png",
+        },
+      ]),
+    ).toBe(
+      [
+        "[[reply_to:message-1]]\nHello\nAttachment: https://example.test/photo.png",
+        "Listen\nAttachment: https://example.test/clip.mp3\n[[audio_as_voice]]",
+        "private",
+      ].join("\n\n"),
+    );
+  });
+});
+
+describe("createChatSendReplyDispatch", () => {
+  it("captures visible replies, promotes tool media, and marks blocked turns", async () => {
+    const markBlocked = vi.fn();
+    const dispatch = createChatSendReplyDispatch({
+      accountId: undefined,
+      isAgentRunStarted: () => false,
+      logGateway: { warn: vi.fn() } as never,
+      session: {
+        agentId: "main",
+        backingSessionId: undefined,
+        cfg: {},
+        clientRunId: "run-1",
+        sessionKey: "agent:main:main",
+        sessionLoadOptions: undefined,
+      },
+      userTurnRecorder: { markBlocked },
+    });
+    const blockedPayload = setReplyPayloadMetadata(
+      { text: "blocked" },
+      { beforeAgentRunBlocked: true },
+    );
+
+    dispatch.dispatcher.sendBlockReply(blockedPayload);
+    dispatch.dispatcher.sendToolResult({
+      text: "tool summary",
+      mediaUrl: "https://example.test/audio.mp3",
+    });
+    dispatch.dispatcher.sendFinalReply({ text: "done" });
+    await dispatch.dispatcher.waitForIdle();
+
+    expect(markBlocked).toHaveBeenCalledOnce();
+    expect(dispatch.deliveredReplies).toEqual([
+      { payload: blockedPayload, kind: "block" },
+      {
+        payload: {
+          text: undefined,
+          mediaUrl: "https://example.test/audio.mp3",
+        },
+        kind: "final",
+      },
+      { payload: { text: "done" }, kind: "final" },
+    ]);
+  });
+});
